* wmi/openvas_smb_interface.h (SMB_HANDLE): Define to long int.
git-svn-id: svn+ssh://scm.wald.intevation.org/openvas/trunk/openvas-smb@27141 423fd1db-d629-0410-8442-d21db03e70f4

+++ wmi/openvas_smb_interface.h
@@ -35,7 +35,7 @@
 #ifndef _NASL_OPENVAS_SMB_INTERFACE_H
 #define _NASL_OPENVAS_SMB_INTERFACE_H
 
-typedef int SMB_HANDLE;
+typedef long int SMB_HANDLE;
 
 char * smb_versioninfo();
 int smb_connect(const char *, const char *, const char *, const char *, SMB_HANDLE *);
